backend/requirements/production.txt
Django~=4.2
djangorestframework
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2-binary
dj-database-url
django-cors-headers
Pillow
